% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_colors.R
\name{ag_colors}
\alias{ag_colors}
\title{ag_colors}
\usage{
ag_colors(palette_name = NULL)
}
\arguments{
\item{palette_name}{A character string specifying the name of the desired color palette. If NULL (default), a palette is randomly selected.}
}
\value{
A vector representing either the selected color palette or a randomly chosen color from a palette.
}
\description{
To define single color for standardizing color use in projects (Source: RColorBrewer)
}
\details{
DEFINE MY COLORS --------------------------------------
}
\examples{
# Select a specific palette
ag_colors("greens")
ag_colors("greens")[4]

# Randomly select a color from any palette
ag_colors()


}
\keyword{color}
\keyword{data}
\keyword{palette,}
\keyword{visualization}
