% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_colors.R
\name{ag_colors_scale}
\alias{ag_colors_scale}
\title{Generate a scaled color palette}
\usage{
ag_colors_scale(color, num_colors = 8, start = 1, end = 0)
}
\arguments{
\item{color}{The color to create the palette from. Options include "gray", "red", "blue", "green", "orange", "cyan", "magenta", "yellow".}

\item{num_colors}{The number of colors in the palette. Default is 8.}

\item{start}{The starting value for the color scale. Default is 1. Used for gray scale.}

\item{end}{The ending value for the color scale. Default is 0. Used for gray scale.}
}
\value{
A vector of colors representing the scaled color palette.
}
\description{
This function generates a scaled color palette based on the specified color.
}
\examples{
\dontrun{
ag_colors_scale("yellow", 1)
ag_colors_scale("gray", 255)
}

}
\author{
Ali Guner
}
