% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_create_project_folders.R
\name{ag_create_excel}
\alias{ag_create_excel}
\title{Create an Excel File for Variables Name}
\usage{
ag_create_excel(dataset = analysisDataset, excel_path = "Data/Variables.xlsx")
}
\arguments{
\item{dataset}{The dataset to extract column names from. Default is 'analysisDataset'.}

\item{excel_path}{The path and filename for the Excel file. Default is "Data/Variables.xlsx".}
}
\value{
Saves a .xlsx file into the path.
}
\description{
This function creates an Excel file with column names of the main dataset.
}
\examples{
\dontrun{
#' ag_create_excel(dataset = iris)
ag_create_excel(dataset = penguins, excel_path = "MyExcelFile.xlsx")
}
}
\author{
Ali Guner
}
