% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_density_plots.R
\name{ag_density_plots}
\alias{ag_density_plots}
\title{Generate Density Plots for Numeric Variables}
\usage{
ag_density_plots(dataset, strata = NULL, table_vars = NULL)
}
\arguments{
\item{dataset}{The input dataset.}

\item{strata}{(Optional) The name of the strata variable for stratified plots.}

\item{table_vars}{(Optional) The names of the variables to be plotted. If not provided, all numeric variables in the dataset will be plotted.}
}
\value{
A density plot for each numeric variable.
}
\description{
This function generates density plots for numeric variables in a dataset. It can also handle stratified plots if a strata variable is provided.
}
\examples{
\dontrun{
# Generate density plots for numeric variables
ag_density_plots(penguins)

# Generate stratified density plots
ag_density_plots(penguins, strata = "species")
}

}
\author{
Ali Guner
}
