% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_columns}
\alias{ag_ff_columns}
\title{Adjusts columns for variables in ag_ff_summary table}
\usage{
ag_ff_columns(.data, levels = FALSE)
}
\arguments{
\item{.data}{The input ag_ff_summary table.}

\item{levels}{Clean levels, default is FALSE.}
}
\value{
The modified ag_ff_summary table with adjusted column values.
}
\description{
This function adjusts column values in an ag_ff_summary table. Converts parametric values (for Total and Groups columns, whichever present) as Mean ± SD, or Median (Q1- Q3)
Also clean levels when levels argument is TRUE.
}
\examples{
\dontrun{
table_vars_1 <- penguins \%>\%
dplyr::select(-species) \%>\% names()

ag_ff_summary(penguins, "species", table_vars_1) \%>\%
ag_ff_relocate("TGP") \%>\%
ag_ff_columns(levels = TRUE)

table_vars_2 <- iris \%>\%
dplyr::select(-Species) \%>\% names()

ag_ff_summary(iris, "Species", table_vars_2) \%>\%
ag_ff_relocate("TGP") \%>\%
ag_ff_columns()
}


}
\seealso{
\code{\link{ag_ff_summary}}
}
\author{
Ali Guner
}
