% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_glimpse}
\alias{ag_ff_glimpse}
\title{Enhanced ff_glimpse Function with FlexTable Output}
\usage{
ag_ff_glimpse(
  .data,
  strata = NULL,
  table_vars = NULL,
  type = NULL,
  missing = FALSE,
  levels_cut = 10,
  ...
)
}
\arguments{
\item{.data}{The input .data.}

\item{strata}{The variable used for stratification. Default is `NULL`.}

\item{table_vars}{The variables to include in the summary table. Default is `NULL`, which includes all variables in the .data.}

\item{type}{The type of variables to include in the summary table. Valid values are `NULL` (default), "cat" (categorical), or "cont" (continuous).}

\item{missing}{Logical indicating whether to include a category for missing values in stratified evaluation Default is `FALSE`.}

\item{levels_cut}{The maximum number of levels to display for categorical variables. Default is 10.}

\item{...}{Additional arguments to be passed to the `ag_flex()` function.}
}
\value{
A `flextable` object containing the summary table.
}
\description{
This function provides an enhanced version of the ff_glimpse function from the `finalfit` package.
It generates a summary table with additional information such as levels, missing values, quartiles, mean, standard deviation, and more. The output is in the form of a `flextable` object, which allows for easy customization and formatting.
#'
}
\examples{
\dontrun{
library(palmerpenguins)
ag_ff_glimpse(penguins, type = "cont")
ag_ff_glimpse(penguins, strata = "sex", missing = TRUE, type = "cat")
}

}
\author{
Ali Guner
}
