% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_labels}
\alias{ag_ff_labels}
\title{Rename labels of a ag_ff_summary table}
\usage{
ag_ff_labels(
  .data,
  use_vector = TRUE,
  vector_name = NULL,
  use_excel = FALSE,
  excel_path = "_Data/Variables.xlsx",
  excel_old_names = "original",
  excel_new_names = "corrected"
)
}
\arguments{
\item{.data}{The ag_ff_summary table to be modified.}

\item{use_vector}{Logical value indicating whether to use a vector for renaming labels. Default is \code{TRUE}.}

\item{vector_name}{The name of the vector containing old and new label names. Required if \code{use_vector = TRUE}.}

\item{use_excel}{Logical value indicating whether to use an Excel file for renaming labels. Default is \code{FALSE}.}

\item{excel_path}{The path to the Excel file containing old and new label names. Required if \code{use_excel = TRUE}.}

\item{excel_old_names}{The name of the column in the Excel file containing the old label names. Default is \code{"original"}.}

\item{excel_new_names}{The name of the column in the Excel file containing the new label names. Default is \code{"corrected"}.}
}
\value{
The ag_ff_summary table with renamed labels.
}
\description{
This function renames labels in ag_ff_summary table based on a vector of old and new names or an Excel file.
You can define old names and new names inside the R chunk. But using external excel file will be more convenient.
If you don't have a Variables.xlsx file, consider using \code{ag_create_excel}
}
\examples{
\dontrun{
label_names <- c(
  "Sepal.Length" = "Sepal length",
  "Sepal.Width" = "Sepal width",
  "Petal.Length" = "Petal length",
  "Petal.Width" = "Petal width",
  "Species" = "Species"
)

ag_ff_summary(iris, "Species", table_vars_2) \%>\%
  ag_ff_relocate("TGP") \%>\%
  ag_ff_columns(levels = TRUE) \%>\%
  ag_ff_labels(use_vector = TRUE, vector_name = label_names)

ag_ff_summary(iris, "Species", table_vars_2) \%>\%
  ag_ff_relocate("TGP") \%>\%
  ag_ff_columns(levels = TRUE) \%>\%
  ag_ff_labels(use_excel = TRUE)
}

}
