% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_non_param_vars}
\alias{ag_ff_non_param_vars}
\title{Extract indices for Non-Normally Distributed Variables}
\usage{
ag_ff_non_param_vars(dataset, table_vars, strata = NULL)
}
\arguments{
\item{dataset}{A data frame containing the variables of interest.}

\item{table_vars}{A character vector of variable names to be tested for normality.}

\item{strata}{An optional grouping variable for stratified analysis. Default is NULL.}
}
\value{
A vector containing the indices of non-normally distributed variables.
}
\description{
This function serves as a helper for finalfit functions, extracting the indices of non-normally
distributed variables based on the Shapiro-Wilk test for normality. It utilizes the
`ag_shapiro` function with specific parameters. Good to use as cont_nonpara argument in summary_factorlist().
names and message arguments are always FALSE.
}
\examples{
\dontrun{
#' table_vars <- c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
or
table_vars <- iris \%>\%
dplyr::select(-Species) \%>\%
names()
dependent <- "Species"

ag_ff_non_param_vars(dataset = iris,
table_vars = table_vars,
strata = dependent)
}
}
\author{
Ali Guner
}
