% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_pull_regression}
\alias{ag_ff_pull_regression}
\title{Extracts a value from an ag_ff_regression table}
\usage{
ag_ff_pull_regression(
  dataset,
  label_value,
  levels_value,
  target,
  label_col = "label",
  levels_col = "levels",
  format = NULL
)
}
\arguments{
\item{dataset}{The input data frame or table.}

\item{label_value}{The value of the label column to filter.}

\item{levels_value}{The value of the levels column to filter.}

\item{target}{The column name of the target value to extract.}

\item{label_col}{The column name for the label column (default is "label").}

\item{levels_col}{The column name for the levels column (default is "levels").}

\item{format}{The desired output format of the extracted value. Options are NULL (default), "bracket", or "parenthesis".}
}
\value{
The extracted value based on the specified criteria and format.
}
\description{
This function extracts a value from an ag_ff_regression table based on specified criteria.
}
\examples{
\dontrun{
ag_ff_pull_regression(dataset, "label_value", "levels_value", "target")
}
}
\author{
Ali Guner
}
