% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_pull_summary}
\alias{ag_ff_pull_summary}
\title{Pull Summary Value from ag_ff_summary Output (to use as inline text)}
\usage{
ag_ff_pull_summary(
  dataset,
  label_value,
  levels_value = NULL,
  target_value,
  label_col = "label",
  levels_col = "levels",
  ff_column = TRUE,
  ff_column_levels = FALSE,
  use_percentage = TRUE,
  use_bracket = FALSE
)
}
\arguments{
\item{dataset}{The dataset to extract the summary value from.}

\item{label_value}{The value of the label variable to filter on.}

\item{levels_value}{The value of the levels variable to filter on (default is \code{NULL}).}

\item{target_value}{The name of the variable to extract the summary value from.}

\item{label_col}{The name of the label column in the dataset (default is "label").}

\item{levels_col}{The name of the levels column in the dataset (default is "levels").}

\item{ff_column}{Logical indicating whether to use fast factor columns (default is \code{TRUE}).}

\item{ff_column_levels}{Logical indicating whether to include levels in fast factor columns (default is \code{FALSE}).}

\item{use_percentage}{Logical indicating whether to format the value as a percentage (default is \code{TRUE}).}

\item{use_bracket}{Logical indicating whether to replace parentheses with square brackets (default is \code{FALSE}).}
}
\value{
The summary value matching the label and levels conditions.
}
\description{
This function is used to pull a summary value from the output of the \code{ag_ff_summary} function.
}
\examples{
\dontrun{
summary_penguins <- ag_ff_summary(palmerpenguins::penguins, strata = "species", table_vars = palmerpenguins::penguins \%>\% select(-species) \%>\% names())
ag_ff_pull_summary(summary_penguins, label_value = "island", levels_value = "Biscoe",  target_value = "Adelie")
}

}
\author{
Ali Guner
}
