% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_regression}
\alias{ag_ff_regression}
\title{Perform regression analysis (including reduced MV) using finalfit::finalfit}
\usage{
ag_ff_regression(
  dataset,
  dependent_var,
  regression_vars = NULL,
  test_var = NULL,
  p_mv = 0.1,
  must_vars_mv = NULL,
  keep_models = FALSE,
  ...
)
}
\arguments{
\item{dataset}{The analysis dataset.}

\item{dependent_var}{The dependent variable. For logistic regression, it should be a binary variable. For survival analysis, it should be a survival object (e.g., "Surv(time_to_event, event_indicator)").}

\item{regression_vars}{A vector of variable names to include as regression variables.}

\item{test_var}{A vector of variable names to include as a test variable.}

\item{p_mv}{The p-value threshold for variable selection. Variables with p-values less than this threshold will be considered significant.}

\item{must_vars_mv}{A vector of variable names that must be included regardless of their p-values.}

\item{keep_models}{Logical value indicating whether to keep the intermediate models generated during the analysis.}

\item{...}{Additional arguments to be passed to the `finalfit::finalfit` function.}
}
\value{
Returns the finalfit analysis output.
}
\description{
This function performs the regression analysis on the specified dataset using the provided variables. It selects significant multivariable variables using the `select_vars_mv` function and then generates the finalfit output using the `finalfit::finalfit` function.
}
\examples{
\dontrun{
dependent_var <- "Surv(time_to_survival, death_alive_numeric)"

regression_output <- ag_ff_regression(analysis_dataset, dependent_var, regression_vars, test_var, p_mv = 0.1, must_vars_mv)
regression_output
}
}
\author{
Ali Guner
}
