% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_regression_simple}
\alias{ag_ff_regression_simple}
\title{Uni- and multivariable regression analysis using finalfit::finalfit package}
\usage{
ag_ff_regression_simple(
  dataset,
  dependent_var,
  regression_vars = NULL,
  test_var = NULL,
  ...
)
}
\arguments{
\item{dataset}{The analysis dataset.}

\item{dependent_var}{The dependent variable. For logistic regression, it should be a binary variable ("event"). For survival analysis, it should be a survival object (e.g., "Surv(time_to_event, event_)").}

\item{regression_vars}{A vector of variable names to include as regression variables.}

\item{test_var}{A vector of variable names to include as a test variable (can be inside regression_vars).}

\item{...}{Additional arguments to be passed to the finalfit function.}
}
\value{
Returns the output of regression tests: label, levels, data summary, UV (OR or HR), MV (OR or HR)
}
\description{
This function performs uni- and multivariable regression tests using the finalfit package. It prepares the dataset, specifies the dependent variable, regression variables, and test variable.
}
\examples{
\dontrun{
dependent_var <- "Surv(time_to_survival, death_alive_numeric)"


ag_ff_regression_simple(analysis_dataset, dependent_var,
regression_vars, test_var, must_vars)
}
}
\author{
Ali Guner
}
