% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_relocate}
\alias{ag_ff_relocate}
\title{Relocate columns of ag_ff_summary table}
\usage{
ag_ff_relocate(.data, order = NULL)
}
\arguments{
\item{.data}{The input (ag_ff_summary table) to be modified.}

\item{order}{Optional. The order to relocate the columns. Default is "GTP".
Possible values are: "G" (Groups), "GP" (Groups and P-value),
"GT" (Groups and Total), "GTP" (Groups, Total, and P-value),
"TGP" (Total, Groups, and P-value), "T" (Total). If the dataset
has only 3 columns, which means no strata was used, the order argument is not needed.}
}
\value{
Modified dataset with relocated columns.
}
\description{
This function relocates columns of finalfit summary table based on the specified order,
considering different strata analysis options. It is primarily designed to
work with datasets generated by the \code{ag_ff_summary} function.
}
\examples{
\dontrun{
table_vars_1 <- penguins \%>\%
dplyr::select(-species) \%>\%
names()

ag_ff_summary(penguins, table_vars = table_vars_1) \%>\%
ag_ff_relocate()

ag_ff_summary(penguins, "species", table_vars_1) \%>\%
ag_ff_relocate("TGP")
}

}
\seealso{
\code{\link{ag_ff_summary}}
}
\author{
Ali Guner
}
