% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_summary_factorlist}
\alias{ag_ff_summary_factorlist}
\title{Aggregated Summary Factor List (extras for finalfit::summary_factorlist)}
\usage{
ag_ff_summary_factorlist(
  dataset,
  dependent,
  table_vars,
  na_include = TRUE,
  row_col_sums = "row_col_based",
  all_cont = "shapiro",
  ...
)
}
\arguments{
\item{dataset}{A data frame containing the variables of interest.}

\item{dependent}{The dependent variable (outcome) in the analysis, should be a factor.}

\item{table_vars}{A character vector specifying the explanatory variables (predictors) for the analysis.
These can be factors, characters, or numerics.}

\item{na_include}{Logical, indicating whether missing values should be included in the analysis.
Default is \code{TRUE}.}

\item{row_col_sums}{A character string specifying the type of summary statistics to be generated.
Options include "row_based" for row-based summaries, "col_based" for column-based summaries,
and "row_col_based" for both row and column-based summaries. Default is "row_col_based".}

\item{all_cont}{A character string specifying the approach for continuous variables. Options include
"parametric" for parametric tests (mean), "nonparametric" for non-parametric tests (median),
and "shapiro" for using the Shapiro-Wilk test for normality and choosing between parametric
and non-parametric accordingly. Default is "shapiro".}

\item{...}{Additional arguments to be passed to the \code{\link{finalfit::summary_factorlist}} function.}
}
\value{
A data frame containing the aggregated summary statistics and test results.
}
\description{
Generate aggregated summary statistics and tests for categorical and continuous variables.

This function generates summary statistics and statistical tests for both categorical and continuous
variables in a dataset. It is designed to work with the `finalfit` package and provides flexibility
in specifying the type of summary statistics and tests to be performed.
}
\examples{
\dontrun{
library(palmerpenguins)
dependent <- "species"
table_vars_1 <- penguins \%>\%
dplyr::select(-species) \%>\% names()
ag_ff_summary_factorlist(penguins, dependent, table_vars_1, na_include = TRUE, row_col_sums = "row_col_based", all_cont = "parametric")
}

}
\seealso{
\code{\link{finalfit::summary_factorlist}}, \code{\link{ag_shapiro}}
}
\author{
Ali Guner
}
\keyword{EDA}
\keyword{analysis}
\keyword{dataset}
\keyword{summary}
