% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_ff_summary_factorlist_corrected}
\alias{ag_ff_summary_factorlist_corrected}
\title{Aggregated Summary Factor List with Fisher Correction}
\usage{
ag_ff_summary_factorlist_corrected(
  dataset,
  dependent,
  table_vars,
  na_include = TRUE,
  row_col_sums = "row_col_based",
  all_cont = "shapiro",
  fisher_correction = TRUE,
  ...
)
}
\arguments{
\item{dataset}{A data frame containing the variables of interest.}

\item{dependent}{The dependent variable (outcome) in the analysis, should be a factor.}

\item{table_vars}{A character vector specifying the explanatory variables (predictors) for the analysis.
These can be factors, characters, or numerics.}

\item{na_include}{Logical, indicating whether missing values should be included in the analysis.
Default is \code{TRUE}.}

\item{row_col_sums}{A character string specifying the type of summary statistics to be generated.
Options include "row_based" for row-based summaries, "col_based" for column-based summaries,
and "row_col_based" for both row and column-based summaries. Default is "row_col_based".}

\item{all_cont}{A character string specifying the approach for continuous variables. Options include
"parametric" for parametric tests (mean), "nonparametric" for non-parametric tests (median),
and "shapiro" for using the Shapiro-Wilk test for normality and choosing between parametric
and non-parametric accordingly. Default is "shapiro".}

\item{fisher_correction}{Logical, indicating whether Fisher's correction should be applied for expected cell
frequencies less than 5 in more than 20% of the cells in a contingency table.
Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed to \code{ag_ff_summary_factorlist}.}
}
\value{
A data frame containing the aggregated summary statistics and test results, with or without Fisher's correction.
}
\description{
Generate aggregated summary statistics and tests for categorical and continuous variables,
with an option for Fisher's correction in contingency tables.

This function extends the functionality of the `ag_ff_summary_factorlist` function by introducing
Fisher's correction for expected cell frequencies less than 5 in more than 20% of the cells in a
contingency table. The function provides flexibility in specifying the type of summary statistics
and tests to be performed, including the option for Fisher's correction.
}
\examples{
\dontrun{
  # Example usage:
  result <- ag_ff_summary_factorlist_corrected(penguins, dependent = "species",
                                                table_vars = c("bill_length_mm", "bill_depth_mm"),
                                                na_include = FALSE, row_col_sums = "row_col_based",
                                                all_cont = "shapiro", fisher_correction = TRUE)
}

}
\seealso{
\code{\link{ag_ff_summary_factorlist}}, \code{\link{ag_fisher}}
}
\author{
Ali Guner
}
\keyword{EDA}
