% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\name{ag_flex_center}
\alias{ag_flex_center}
\title{Center align the content of a flextable}
\usage{
ag_flex_center(flex_obj)
}
\arguments{
\item{flex_obj}{The flextable object for which the content alignment is applied.}
}
\value{
The flextable object with the content aligned to the center.
}
\description{
This function center aligns the content of a flextable starting from the third column to the last column.
}
\examples{
\dontrun{
ag_flex_center(flex_obj)
}
}
\seealso{
\code{\link{ag_flex()}}
}
\author{
Ali Guner
}
