% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\name{ag_flex_footnote}
\alias{ag_flex_footnote}
\title{Add a footnote to a flextable}
\usage{
ag_flex_footnote(flex_obj, footnote = NULL)
}
\arguments{
\item{flex_obj}{The flextable object to which the footnote is added.}

\item{footnote}{The text of the footnote.}
}
\value{
The flextable object with the footnote added.
}
\description{
This function adds a footnote to a flextable object.
}
\examples{
\dontrun{
ag_flex_footnote(flex_obj, "This is a footnote.")
}

}
\seealso{
\code{\link{ag_flex()}}
}
\author{
Ali Guner
}
