% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\name{ag_flex_header}
\alias{ag_flex_header}
\title{Generate and set header labels for an aggregated flextable and finalfit}
\usage{
ag_flex_header(flex_obj, .dataset, strata)
}
\arguments{
\item{flex_obj}{The flextable object for which the header labels are generated and set.}

\item{.dataset}{The dataset for which the header labels are generated.}

\item{strata}{The strata variable used for aggregation.}
}
\value{
The flextable object with the header labels set.
}
\description{
This function generates header labels for an aggregated flextable based on the provided dataset and strata variable,
and sets the labels in the flextable object.
}
\examples{
\dontrun{
ag_flex_header(flex_obj, dataset, "strata_variable")
}

}
\seealso{
\code{\link{ag_flex()}}
}
\author{
Ali Guner
}
