% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\name{ag_flex_header_labels}
\alias{ag_flex_header_labels}
\title{Add Custom Headers (with n) to a flextable Object}
\usage{
ag_flex_header_labels(
  flex_obj,
  .dataset,
  strata,
  custom = FALSE,
  custom_levels
)
}
\arguments{
\item{flex_obj}{A flextable object to which headers will be added.}

\item{.dataset}{The dataset used for generating the headers.}

\item{strata}{The name of the variable used for grouping or stratifying the data.}

\item{custom}{Logical value indicating whether to use custom levels or not.}

\item{custom_levels}{A named vector specifying custom levels for the strata variable.}
}
\value{
A flextable object with custom headers added.
}
\description{
This function adds custom headers to a flextable object based on specified
dataset, strata variable, and custom levels.
}
\examples{
\dontrun{

custom_levels <- c("Gentoo" = "MyGentoo",
"Adelie" = "MyAdelie")

ag_ff_summary(penguins, "species", penguins \%>\% select(-species) \%>\% names()) \%>\%
        ag_flex() \%>\%
        ag_flex_header_labels(penguins, strata = "species", custom = TRUE, custom_levels = custom_levels)

}
}
\author{
Ali Guner
}
