% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\name{ag_flex_hline}
\alias{ag_flex_hline}
\title{Add horizontal lines to a flextable}
\usage{
ag_flex_hline(flex_obj)
}
\arguments{
\item{flex_obj}{The flextable object to which the horizontal lines are added.}
}
\value{
The flextable object with horizontal lines added.
}
\description{
This function adds horizontal lines to a flextable based on the non-empty label rows in the dataset.
}
\examples{
\dontrun{
ag_flex_hline(flex_obj)
}

}
\seealso{
\code{\link{ag_flex()}}
}
\author{
Ali Guner
}
