% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\name{ag_flex_save}
\alias{ag_flex_save}
\title{Save a flextable as a Word document}
\usage{
ag_flex_save(
  flex_obj,
  n = "x",
  orientation = "Landscape",
  mypath = "_Outputs/_Tables"
)
}
\arguments{
\item{flex_obj}{The flextable object to be saved.}

\item{n}{The table number or identifier. Default is "x".}

\item{orientation}{The orientation of the page. Can be "Landscape" or "Portrait". Default is "Landscape".}

\item{mypath}{The path where the Word document will be saved. Default is "_Outputs/_Tables" in the current working directory.}
}
\value{
The saved Word document.
}
\description{
This function saves a flextable object as a Word document.
}
\examples{
\dontrun{
ag_flex_save(flex_obj, n = 1, orientation = "Landscape", mypath = "path/to/save")
}

}
\seealso{
\code{\link{flextable::save_as_docx()}}
}
\author{
Ali Guner
}
