% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\name{ag_flex_title}
\alias{ag_flex_title}
\title{Add a title to a flextable}
\usage{
ag_flex_title(flex_obj, n, title)
}
\arguments{
\item{flex_obj}{The flextable object to which the title is added.}

\item{n}{The table number.}

\item{title}{The title text.}
}
\value{
The flextable object with the title added.
}
\description{
This function adds a title to a flextable object.
}
\examples{
\dontrun{
ag_flex_title(flex_obj, 1, "Table Title")
}

}
\seealso{
\code{\link{ag_flex()}}
}
\author{
Ali Guner
}
