% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_generate_template_quarto.R
\name{ag_generate_template_quarto}
\alias{ag_generate_template_quarto}
\title{Generate Quarto Document Using Template}
\usage{
ag_generate_template_quarto(name = "01 initial.qmd", open = interactive(), ...)
}
\arguments{
\item{name}{A character string specifying the name of the output Quarto document.
If not provided, the default name is "01 initial.qmd".}

\item{open}{A logical indicating whether to open the newly created Quarto document interactively.}

\item{...}{Additional arguments passed to the \code{use_template} function.}
}
\value{
None
}
\description{
The \code{ag_generate_template_quarto()} function creates a new Quarto (.qmd) document
using a template file provided by the "MyR" package. This function simplifies the process
of initializing Quarto projects with a customizable template.
}
\examples{
\dontrun{
# Example usage
ag_generate_template_quarto(name = "my_quarto_project.qmd", open = TRUE)
}

}
\seealso{
\code{\link{use_template}}
}
