% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_colors.R
\name{ag_make_colors}
\alias{ag_make_colors}
\title{Generate Dark-Light Colors}
\usage{
ag_make_colors(num_colors, seed = 2023)
}
\arguments{
\item{num_colors}{An integer specifying the number of colors to generate in the color palette.}

\item{seed}{An optional argument specifying the seed for the random number generator. If not provided, the default seed of 2023 will be used.}
}
\value{
A list with the following components (1000 different options were defined):
\describe{
  \item{palette}{A list containing the color palette. Each color in the palette is represented as a list with the following components: \code{normal_color}, \code{darker_color}, and \code{lighter_color}.}
  \item{normal}{A vector of the normal colors in the palette.}
  \item{darker}{A vector of the darker colors in the palette.}
  \item{lighter}{A vector of the lighter colors in the palette.}
}
}
\description{
The \code{ag_make_colors} function generates a color palette consisting of normal, darker, and lighter colors based on random hues.
}
\examples{
\dontrun{
# Generate a color palette with 5 colors using the default seed:
colors <- ag_make_colors(5)

# Generate a color palette with 10 colors using a specific seed:
colors <- ag_make_colors(10, seed = 1967)

# Extract the color palette
palette <- colors$palette

# Extract the normal colors
normal_colors <- colors$normal

# Extract the darker colors
darker_colors <- colors$dark

# Extract the lighter colors
lighter_colors <- colors$light
}

#' @export
}
\author{
Ali Guner
}
\keyword{color}
\keyword{generator}
\keyword{palette}
