% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_shapiro.R
\name{ag_non_param_vars}
\alias{ag_non_param_vars}
\title{Extract indices for Non-Normally Distributed Variables}
\usage{
ag_non_param_vars(.data, strata = NULL, table_vars)
}
\arguments{
\item{.data}{A data frame containing the variables of interest.}

\item{strata}{An optional grouping variable for stratified analysis. Default is NULL.}

\item{table_vars}{A character vector of variable names to be tested for normality.}
}
\value{
A vector containing the indices of non-normally distributed variables.
}
\description{
This function serves as a helper for finalfit functions, extracting the indices of non-normally
distributed variables based on the Shapiro-Wilk test for normality. It utilizes the
`ag_shapiro` function with specific parameters. Good to use as cont_nonpara argument in summary_factorlist().
}
\examples{
\dontrun{
#' table_vars <- c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
or
table_vars <- iris \%>\%
dplyr::select(-Species) \%>\%
names()
dependent <- "Species"

ag_non_param_vars(.data = iris,
strata = dependent,
table_vars = table_vars
)
}
}
\author{
Ali Guner
}
