% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_qq_plots.R
\name{ag_qq_plot_single}
\alias{ag_qq_plot_single}
\title{Generate QQ plots for numeric variables}
\usage{
ag_qq_plot_single(dataset, strata = NULL, table_vars = NULL)
}
\arguments{
\item{dataset}{The input dataset.}

\item{strata}{The variable used for stratification. If NULL, no stratification is performed.}

\item{table_vars}{Optional. The variables to be included in the QQ plots. If NULL, all numeric variables in the dataset are used.}
}
\value{
A list containing the generated plot and the number of missing values in the stratification variable (if applicable).
}
\description{
This function generates QQ plots for numeric variables in a dataset. It allows for stratified analysis based on a specified variable.
}
\examples{
\dontrun{
ag_qq_plot_single(penguins, strata = NULL)$plot

table_vars_1 <- penguins \%>\%
dplyr::select(-species) \%>\% names()
ag_qq_plot_single(penguins, strata = "species", table_vars = table_vars_1)$plot
}

}
\author{
Ali Guner
}
