% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_qq_plots.R
\name{ag_qq_plots}
\alias{ag_qq_plots}
\title{Generate multiple QQ plots for all/selected numeric variables}
\usage{
ag_qq_plots(dataset, strata = NULL, table_vars = NULL, n_row = 5)
}
\arguments{
\item{dataset}{The input dataset.}

\item{strata}{The variable used for stratification. If NULL, no stratification is performed.}

\item{table_vars}{Optional. The variables to be included in the QQ plots. If NULL, all numeric variables in the dataset are used.}

\item{n_row}{The number of plots to display in each plot. Default is 5.}
}
\value{
None (plots are displayed)
}
\description{
This function generates multiple QQ plots for numeric variables in a dataset. It allows for stratified analysis based on a specified variable.
}
\examples{
\dontrun{
ag_qq_plots(penguins)
#' table_vars_1 <- penguins \%>\%
dplyr::select(-species) \%>\% names()

ag_qq_plots(penguins, "sex", table_vars_1)
}

}
\seealso{
[func(ag_qq_plot_single)]
}
\author{
Ali Guner
}
