% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_relabel_excel}
\alias{ag_relabel_excel}
\title{Relabel a column in a dataset using an Excel file as a source of label mappings}
\usage{
ag_relabel_excel(
  .data,
  label_column = "label",
  excel_path,
  sheet_name = "Sheet 1",
  excel_old_names,
  excel_new_names,
  add_units = FALSE
)
}
\arguments{
\item{.data}{The input dataset.}

\item{label_column}{The name of the column to be relabeled. Default is "label".}

\item{excel_path}{The path to the Excel file.}

\item{sheet_name}{The name of the sheet in the Excel file containing the label mappings. Default is "Sheet 1".}

\item{excel_old_names}{The name of the column in the Excel file containing the original labels.}

\item{excel_new_names}{The name of the column in the Excel file containing the new labels.}

\item{add_units}{Logical. If TRUE, adds units information to the new labels based on the "units" column in the Excel file. Default is FALSE.}
}
\value{
A modified dataset with the specified column relabeled.
}
\description{
This function relabels a specified column in a dataset by replacing its values with
corresponding labels from an Excel file. The Excel file should contain two columns:
one with the original labels and another with the new labels.
}
\examples{
\dontrun{
relabeled_data <- ag_relabel_excel(.data = data, excel_path = "path/to/excel.xlsx",
                                   excel_old_names = "original_labels",
                                   excel_new_names = "new_labels",
                                   add_units = TRUE)
}

}
