% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ag_relabel_vector}
\alias{ag_relabel_vector}
\title{Relabel a column using label mappings}
\usage{
ag_relabel_vector(.data, label_column, vector_name)
}
\arguments{
\item{.data}{The input dataset.}

\item{label_column}{The name of the column to be relabeled.}

\item{vector_name}{A named vector containing the label mappings.}
}
\value{
A modified dataset with the specified column relabeled.
}
\description{
This function relabels a specified column in a dataset by replacing its values with corresponding labels from a provided vector of label mappings.
}
\examples{
\dontrun{
new_labels <- c(
        "island" = "Island",
        "flipper_length_mm" = "Flipper length (mm)",
        "bill_depth_mm" = "Bill depth (mm)")

ag_ff_summary(penguins, strata = strata, table_vars = table_vars_penguins) \%>\%
        ag_relabel_vector("label", new_labels)
}

}
\author{
Ali Guner
}
