% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_shapiro.R
\name{ag_shapiro}
\alias{ag_shapiro}
\title{Perform Shapiro-Wilk Test for Normality}
\usage{
ag_shapiro(
  .data,
  strata = NULL,
  table_vars = NULL,
  silence = FALSE,
  names = TRUE
)
}
\arguments{
\item{.data}{A data frame containing the variables of interest.}

\item{strata}{An optional grouping variable for stratified analysis. Default is NULL.}

\item{table_vars}{A character vector of variable names to be tested for normality. Can be defined with names(select()) functions.
If NULL, function will use all numeric variables in .data, except for strata if present (NOT RECOMMENDED,
it is better idea to define table_vars)}

\item{silence}{Logical. If FALSE, displays informative messages about the analysis. Default is FALSE}

\item{names}{Logical. If TRUE, returns the names of non-normally distributed variables;
if FALSE, returns the indices of these variables. Default is TRUE}
}
\value{
A character vector containing the names of non-normally distributed variables or an integer count,
based on the specified value of the 'names' argument.
}
\description{
This function conducts the Shapiro-Wilk test for normality on numeric variables in a .data,
either for the entire .data or within specified strata. It returns the names or counts of
variables that are found to be non-normally distributed based on a significance level of 0.05.
Before this test, all variables which are desired to be tests, should be defined in explanatory vector.
}
\examples{
\dontrun{

table_vars <- c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
or
table_vars <- iris \%>\%
dplyr::select(-Species) \%>\%
names()
dependent <- "Species"

ag_shapiro(.data = iris, table_vars = table_vars)
ag_shapiro(.data = iris, table_vars = table_vars, silence = TRUE, names = FALSE)
}


}
\author{
Ali Guner
}
