% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_shapiro.R
\name{ag_shapiro_results}
\alias{ag_shapiro_results}
\title{Calculate Shapiro-Wilk p-values}
\usage{
ag_shapiro_results(dataset, strata = NULL, table_vars = NULL, asteriks = TRUE)
}
\arguments{
\item{dataset}{The input dataset.}

\item{strata}{(Optional) A column name in the dataset to stratify the analysis by. Default is \code{NULL}.}

\item{table_vars}{A character vector of variable names to include in the analysis.}

\item{asteriks}{if \code{TRUE}, test results will be reader-friendly. Default is \code{TRUE}.}
}
\value{
A tibble with columns for the variable names and Shapiro-Wilk p-values.
}
\description{
This function calculates Shapiro-Wilk p-values for normality testing of numeric variables in a dataset.
}
\examples{
table_vars_1 <- penguins \%>\%
dplyr::select(-species) \%>\% names()

ag_shapiro_results(penguins, "species", table_vars_1)

}
