% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R, R/flextable_functions.R
\docType{data}
\name{col_table_header}
\alias{col_table_header}
\title{Constants for package}
\format{
a vector with the colors I need

An object of class \code{character} of length 1.
}
\usage{
col_table_header

col_table_header
}
\value{
A set of color values for table styling.
}
\description{
This file contains commonly used constants for the package.

This code chunk defines colors for different elements of a table, such as the table header, footer,
border, background of the table header, and the overall table border.
}
\examples{
\dontrun{
# Assuming you want to use these colors in table styling
iris \%>\%
ag_flex() \%>\%
flextable::bg(i = 1, j = NULL, bg = col_table_header, part = "header")
}


}
\author{
Ali Guner
}
\keyword{datasets}
\keyword{internals}
