% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{ff_row_col_sums}
\alias{ff_row_col_sums}
\title{Generate Row-, Column-based or both-based Summary Statistics}
\usage{
ff_row_col_sums(
  dataset,
  strata = NULL,
  table_vars = NULL,
  na_include = NULL,
  cont = NULL,
  cont_nonpara = NULL,
  fisher_correction = NULL,
  row_col_sums,
  ...
)
}
\arguments{
\item{dataset}{A data frame containing the variables of interest.}

\item{strata}{The dependent variable for hypothesis testing.}

\item{table_vars}{A vector of explanatory variables for analysis.}

\item{na_include}{Logical. Should NA values be included in the analysis? Default is NULL.}

\item{cont}{The type of measure for continuous variables. Default is NULL.}

\item{cont_nonpara}{A vector specifying non-parametric measures for continuous variables. Default is NULL.}

\item{fisher_correction}{Logical. Should Fisher's exact tests be corrected? Default is NULL.}

\item{row_col_sums}{The method for generating row or column-based summary statistics. Use 'row_based' or 'col_based'.}

\item{...}{Additional arguments to be passed to the underlying functions.}
}
\value{
A summary factor list containing row or column-based summary statistics.
}
\description{
This function generates row or column-based summary statistics based on the specified method.
}
\examples{
\dontrun{
# Example usage of ff_row_col_sums function
row_col_result <- ff_row_col_sums(my_data, strata = "Response", table_vars = c("Var1", "Var2"), row_col_sums = "row_based")
}
}
\seealso{
\code{\link{short_ff}}
}
\author{
Ali Guner
}
