% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\docType{data}
\name{flex_landscape}
\alias{flex_landscape}
\title{Create Landscape Page Section for Flextable}
\format{
An object of class \code{prop_section} (inherits from \code{prop}) of length 3.
}
\usage{
flex_landscape
}
\arguments{
\item{page_size}{An officer page size object specifying the orientation of the page.}

\item{type}{Character, the type of section to be added ("nextPage" for a new page).}

\item{page_margins}{An officer page margin object defining the margins for the page.}
}
\value{
A landscape-oriented page section for use in creating a Word document with officer.
}
\description{
This function generates a landscape-oriented page section for use in creating a Microsoft Word document with the `officer` package.
The section is defined with landscape orientation, and default page margins are applied.
}
\examples{
\dontrun{
# Create a landscape page section
flex_landscape <- officer::prop_section(
  page_size = officer::page_size(orient = "landscape"),
  type = "nextPage",
  page_margins = officer::page_mar()
)
}

}
\author{
Ali Guner
}
\keyword{datasets}
