% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\name{flex_page_section}
\alias{flex_page_section}
\title{Page section}
\usage{
flex_page_section(orientation)
}
\arguments{
\item{orientation}{A character value specifying the page orientation. Use "landscape" for landscape orientation and "portrait" for portrait orientation.}
}
\value{
A page section object for use in a flextable document (while saving with pr_section ()).
}
\description{
Page section
}
\examples{
\dontrun{
# Assuming you want to create a landscape page section
flex_page_section("landscape")

# Assuming you want to create a portrait page section
flex_page_section("portrait")
}

}
\author{
Ali Guner
}
