% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\docType{data}
\name{flex_portrait}
\alias{flex_portrait}
\title{Create Portrait Page Section for Flextable}
\format{
An object of class \code{prop_section} (inherits from \code{prop}) of length 3.
}
\usage{
flex_portrait
}
\arguments{
\item{page_size}{An officer page size object specifying the orientation and width of the page.}

\item{type}{Character, the type of section to be added ("nextPage" for a new page).}

\item{page_margins}{An officer page margin object defining the margins for the page.}
}
\value{
A portrait-oriented page section for use in creating a Word document with officer.
}
\description{
This function generates a portrait-oriented page section for use in creating a Microsoft Word document with the `officer` package.
The section is defined with portrait orientation and a specified width, and default page margins are applied.
}
\examples{
\dontrun{
# Create a portrait page section
flex_portrait <- officer::prop_section(
  page_size = officer::page_size(orient = "portrait", width = 14),
  type = "nextPage",
  page_margins = officer::page_mar()
)
}
}
\author{
Ali Guner
}
\keyword{datasets}
