% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_colors.R
\name{karadeniz}
\alias{karadeniz}
\title{karadeniz color pallette}
\usage{
karadeniz(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose your favourite palette: "Hidirnebi", "Kadirga", "Kayabasi",  "Erikbeli", "Pokut", "Samistal", "Elevit", "Gito" , "Ovit", "Vazil"  or "Palovit"}

\item{reverse}{TRUE/FALSE}

\item{...}{Additional arguments to be passed}
}
\description{
Multiple color palettes based on the "palette" argument
}
\examples{
\dontrun{
karadeniz("Vazil")
colors
show_col(karadeniz("Vazil")(12))
}


}
\author{
Ali Guner
}
\keyword{internal}
