% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\docType{data}
\name{my_blues}
\alias{my_blues}
\title{blues
#' Vector of blues}
\format{
An object of class \code{character} of length 8.
}
\usage{
my_blues
}
\description{
To define my blues palette to standardize color use in projects
}
\examples{
my_blues[4]

}
\author{
Ali Guner
}
\keyword{datasets}
