% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\docType{data}
\name{my_gradient}
\alias{my_gradient}
\title{gradient}
\format{
An object of class \code{character} of length 4.
}
\usage{
my_gradient
}
\description{
To define my gradient palette to standardize color use in projects
}
\examples{
my_gradient[4]

}
\author{
Ali Guner
Vector of gradient
}
\keyword{datasets}
