% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\docType{data}
\name{my_grays}
\alias{my_grays}
\title{grays}
\format{
An object of class \code{character} of length 8.
}
\usage{
my_grays
}
\description{
To define my grays palette to standardize color use in projects
}
\examples{
my_grays[4]

}
\author{
Ali Guner
Vector of grays
}
\keyword{datasets}
