% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\docType{data}
\name{my_greens}
\alias{my_greens}
\title{greens}
\format{
An object of class \code{character} of length 8.
}
\usage{
my_greens
}
\description{
To define my greens palette to standardize color use in projects
}
\examples{
my_greens[4]

}
\author{
Ali Guner
Vector of greens
}
\keyword{datasets}
