% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\docType{data}
\name{my_oranges}
\alias{my_oranges}
\title{oranges}
\format{
An object of class \code{character} of length 8.
}
\usage{
my_oranges
}
\description{
To define my oranges palette to standardize color use in projects
}
\examples{
my_oranges[4]

}
\author{
Ali Guner
Vector of oranges
}
\keyword{datasets}
