% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\docType{data}
\name{my_reds}
\alias{my_reds}
\title{reds}
\format{
An object of class \code{character} of length 8.
}
\usage{
my_reds
}
\description{
To define my reds palette to standardize color use in projects
}
\examples{
my_reds[4]

}
\author{
Ali Guner
Vector of reds
}
\keyword{datasets}
