% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_colors.R
\name{scale_color_karadeniz}
\alias{scale_color_karadeniz}
\title{scale_color_karadeniz}
\usage{
scale_color_karadeniz(
  palette = "Hidirnebi",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{hoose your favourite palette: "Hidirnebi", "Kadirga", "Kayabasi",  "Cal", "Pokut", "Samistal", "Elevit", "Gito" , "Ovit", "Vazil"  or "Palovit"}

\item{discrete}{TRUE/FALSE}

\item{reverse}{TRUE/FALSE}
}
\description{
to COLOR function
}
\examples{
ggplot(data=iris,mapping=aes(x=Petal.Length, y=Petal.Width, color=Species))+
geom_point() + scale_color_karadeniz()

}
\author{
Ali Guner
}
