% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_colors.R
\name{scale_fill_karadeniz}
\alias{scale_fill_karadeniz}
\title{scale_fill_karadeniz}
\usage{
scale_fill_karadeniz(palette = "Faroz", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose your favorite palette: "Hidirnebi", "Kadirga", "Kayabasi", "Cal", "Pokut", "Samistal", "Elevit", "Gito", "Ovit", "Palovit", "Faroz", "Vazil"}

\item{discrete}{TRUE/FALSE}

\item{reverse}{TRUE/FALSE}
}
\description{
To define my color pallette to standardize color use in projects
}
\examples{
\dontrun{
#' iris \%>\%
ggplot(aes(Petal.Length, Petal.Width, fill = Species))+
geom_point() +
scale_fill_karadeniz()
}

iris \%>\%
ggplot(aes(Sepal.Length, Sepal.Width, fill = Petal.Length)) +
       geom_point() +
       scale_fill_karadeniz(palette = "Vazil", discrete = FALSE)
}
\author{
Ali Guner
}
