% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{select_vars_mv}
\alias{select_vars_mv}
\title{Select significant variables from univariable analysis to use in multivariable analysis}
\usage{
select_vars_mv(
  dataset,
  dependent_var,
  regression_vars = NULL,
  test_var = NULL,
  p_mv = 0.1,
  must_vars_mv = NULL,
  ...
)
}
\arguments{
\item{dataset}{The analysis dataset.}

\item{dependent_var}{The dependent variable. For logistic regression, it should be a binary variable. For survival analysis, it should be a survival object (e.g., "Surv(time_to_event, event_indicator)").}

\item{regression_vars}{A vector of variable names to include as regression variables.}

\item{test_var}{A vector of variable names to include as a test variable.}

\item{p_mv}{The p-value threshold for variable selection. Variables with p-values less than this threshold will be considered significant.}

\item{must_vars_mv}{A vector of variable names that must be included regardless of their p-values.}

\item{...}{Additional arguments to be passed to the `ag_ff_regression_simple` function.}
}
\value{
Returns a vector of selected variables that are significant based on the defined p-value threshold and the must-include variables.
}
\description{
This function selects significant variables from a multivariable analysis based on a defined p-value threshold. It uses the `ag_ff_regression_simple` function to perform the initial analysis and then filters the variables based on the specified criteria.
}
\examples{
\dontrun{
dependent_var <- "Surv(time_to_survival, death_alive_numeric)"

must_vars_mv <- c("sarcopenia_v02_simple")

selected_vars <- select_vars_mv(analysis_dataset, dependent_var, regression_vars, test_var, p_mv = 0.1, must_vars_mv)
selected_vars
}
}
\keyword{internal}
