% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalfit_functions.R
\name{short_ff}
\alias{short_ff}
\title{Minimized version of summary_factorlist}
\usage{
short_ff(
  .dataset,
  strata = NULL,
  table_vars = NULL,
  na_include = NULL,
  cont = NULL,
  cont_nonpara = NULL,
  fisher_correction = fisher_correction,
  ...
)
}
\arguments{
\item{.dataset}{A data frame containing the variables of interest.}

\item{strata}{The dependent variable for hypothesis testing.}

\item{table_vars}{A vector of explanatory variables for analysis.}

\item{na_include}{Logical. Should NA values be included in the analysis? Default is NULL.}

\item{cont}{The type of measure for continuous variables. Default is NULL.}

\item{cont_nonpara}{A vector specifying non-parametric measures for continuous variables. Default is NULL.}

\item{fisher_correction}{Logical. Should Fisher's exact tests be corrected? Default is NULL.}

\item{...}{Additional arguments to be passed to the underlying functions.}
}
\value{
A summary factor list containing short summary statistics and hypothesis testing results.
}
\description{
Perform Short Summary Statistics and Hypothesis Testing

This function generates short summary statistics and conducts hypothesis testing, including chi-square tests and Fisher's exact tests if specified.
}
\examples{
\dontrun{
# Example usage of short_ff function
short_result <- short_ff(my_data, strata = "Response", table_vars = c("Var1", "Var2"))
}
}
\seealso{
\code{\link{summary_factorlist}}, \code{\link{ag_fisher}}
}
\author{
Ali Guner
}
\keyword{internal}
