% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_colors.R
\name{to_select_colors}
\alias{to_select_colors}
\title{to_select_colors}
\usage{
to_select_colors(...)
}
\description{
To define my color pallette to standardize color use in projects
}
\examples{
\dontrun{
# Example usage
to_select_colors("blue")
}
}
\author{
Ali Guner
}
