% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_create_project_folders.R
\name{ag_create_project_folders}
\alias{ag_create_project_folders}
\title{Create Project Folders}
\usage{
ag_create_project_folders()
}
\description{
This function automates the creation of essential folder structures for a new R project.
It sets up main directories such as "_Codes," "_Outputs," and "_Data," along with subdirectories,
subfolders, and necessary files for organizing the project efficiently.
The function performs the following steps:

1. Creates main folders: "_Codes," "_Outputs," and "_Data."
2. Sets up subfolders within "_Codes," including "_Functions."
3. Creates a README file within "_Codes" named "00 README.md" for project documentation.
4. Sets up subfolders within "_Outputs," including "_Figures" and "_Tables."
5. Sets up subfolders within "_Data," including "_Raw" and "_Processed."
6. Checks for and deletes a single .qmd file in the main directory with user confirmation.
7. Inserts a new .qmd template file named "01 initial.qmd" into the "_Codes" folder.

This function is designed to be used right after creating a new R project via "New Project" or any other project creation method.
}
\examples{
\dontrun{
# Example usage
ag_create_project_folders()
}

}
\seealso{
\code{\link{dir.create}}, \code{\link{file.remove}}, \code{\link{writeLines}}, \code{\link{ag_generate_template_quarto}}
}
\keyword{folders}
\keyword{initialization}
\keyword{organization}
\keyword{project}
