% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_create_project_folders.R
\name{ag_create_project_folders}
\alias{ag_create_project_folders}
\title{Create Project Folders}
\usage{
ag_create_project_folders()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
This function creates the essential folder structure for a new R project.
It includes the main folders, subfolders, and necessary files to organize the project.
Should be used just after "the project was created" via create_project() or "New Project"
}
\details{
The function performs the following steps:
1. Define folder names, including "_Codes," "_Outputs," and "_Data."
2. Check and create the main folders: "_Codes," "_Outputs," and "_Data."
3. Check and create a subfolder within "_Codes" named "_Functions."
4. Check and create a README file within "_Codes" named "README.md" for project documentation.
5. Find .qmd files in the main directory, and if only one is found, move and rename it to
"_Codes" as "00 initial.qmd" (with a warning if there are multiple .qmd files).
6. Check and create subfolders within "_Outputs," including "_Figures" and "_Tables."
7. Check and create subfolders within "_Data," including "_Raw" and "_Processed."
}
\examples{
\dontrun{
# Example usage
ag_create_project_folders()
}

}
\seealso{
\code{\link{dir.create}}, \code{\link{file.rename}}, \code{\link{writeLines}}
}
\keyword{folders}
\keyword{initialization}
\keyword{organization}
\keyword{project}
