% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_fisher.R
\name{ag_fisher}
\alias{ag_fisher}
\title{Identify Factors with Low Expected Cell Frequencies (for Fisher test)}
\usage{
ag_fisher(
  .dataset,
  strata,
  table_vars,
  silence = TRUE,
  observed_tables = FALSE,
  expected_tables = FALSE
)
}
\arguments{
\item{.dataset}{The .dataset containing the variables of interest.}

\item{strata}{The stratifying factor variable for the analysis.}

\item{table_vars}{A character vector of factor variable names to analyze.}

\item{silence}{Logical, indicating whether to suppress warnings. Default is TRUE.}

\item{observed_tables}{Logical, indicating whether to display contingency tables for observed values. Default is FALSE.}

\item{expected_tables}{Logical, indicating whether to display tables for expected values. Default is FALSE.}
}
\value{
A character vector of factor variable names with low expected cell frequencies, or NULL if none are found.
}
\description{
This function performs a contingency table analysis on specified factor or character variables in a given .dataset,
with an option to identify variables exhibiting expected cell frequencies less than 5 in more than 20% of cells.
It is designed to handle potential issues gracefully, providing informative warnings and excluding rows with
missing values in the strata variable when necessary. The function can display observed and/or expected values
tables based on user preferences.
}
\examples{
\dontrun{
# Usage example:
library(palmerpenguins)

table_vars <- penguins \%>\%
dplyr::select(-species) \%>\%
names()

strata <- "species"

ag_fisher(penguins, strata, table_vars,  silence = TRUE, observed_tables = TRUE)
ag_fisher(penguins, strata, table_vars,   expected_tables = TRUE)
}

}
\seealso{
\code{\link{chisq.test}}
}
\author{
Ali Guner
}
\keyword{chi-square}
\keyword{contingency-tables}
\keyword{fisher}
