% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_functions.R
\name{ag_flex}
\alias{ag_flex}
\title{Create a Styled Flextable}
\usage{
ag_flex(data, flex_font_family = "Arial", flex_font_size = 14)
}
\arguments{
\item{data}{A data frame or a tibble to be displayed in the flextable.}

\item{flex_font_family}{Character, font family for the text in the flextable.}

\item{flex_font_size}{Numeric, font size for the text in the flextable.}
}
\value{
A flextable object with customized styling.
}
\description{
This function generates a styled flextable with specified font family and font size.
The flextable is configured with default font settings, the 'theme_booktabs', and additional styling to make the first row bold.
}
\examples{
\dontrun{
# Assuming you have a data frame named 'my_data'
my_data \%>\% ag_flex()
my_data \%>\% ag_flex(flex_font_family = "Times New Roman", flex_font_size = 14)

# If you want to undo bold()
bold(i = 1, j = NULL, bold = FALSE, part = "header")
}

}
\author{
Ali Guner
}
