% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_colors.R
\name{ag_make_colors}
\alias{ag_make_colors}
\title{Generate Dark-Light Colors}
\usage{
ag_make_colors(num_colors, seed = 2023)
}
\arguments{
\item{num_colors}{An integer specifying the number of colors to generate in the color palette.}

\item{seed}{An optional argument specifying the seed for the random number generator. If not provided, the default seed of 2023 will be used.}
}
\value{
A list with the following components:
\describe{
  \item{palette}{A list containing the color palette. Each color in the palette is represented as a list with the following components: \code{normal_color}, \code{darker_color}, and \code{lighter_color}.}
  \item{normal}{A vector of the normal colors in the palette.}
  \item{darker}{A vector of the darker colors in the palette.}
  \item{lighter}{A vector of the lighter colors in the palette.}
}
}
\description{
The \code{ag_make_colors} function generates a color palette consisting of normal, darker, and lighter colors based on random hues.
}
\examples{
\dontrun{
# Generate a color palette with 5 colors using the default seed:
colors <- ag_make_colors(5)

# Generate a color palette with 10 colors using a specific seed (e.g., 123):
colors <- ag_make_colors(10, seed = 123)

ag_make_colors(4)$palette

num_colors <- 8
ag_make_colors(num_colors)$lighter
ag_make_colors(num_colors)$darker
}

}
\author{
Ali Guner
}
