% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_shapiro.R
\name{ag_shapiro}
\alias{ag_shapiro}
\title{Perform Shapiro-Wilk Test for Normality}
\usage{
ag_shapiro(dataset, strata = NULL, table_vars, silence = TRUE, names = FALSE)
}
\arguments{
\item{dataset}{A data frame containing the variables of interest.}

\item{strata}{An optional grouping variable for stratified analysis. Default is NULL.}

\item{table_vars}{A character vector of variable names to be tested for normality. Can be defined with names(select()) functions.}

\item{silence}{Logical. If FALSE, displays informative messages about the analysis. Default is TRUE.}

\item{names}{Logical. If TRUE, returns the names of non-normally distributed variables;
if FALSE, returns the indices of these variables. Default is FALSE.}
}
\value{
A character vector containing the names of non-normally distributed variables or an integer count,
based on the specified value of the 'names' argument.
}
\description{
This function conducts the Shapiro-Wilk test for normality on numeric variables in a dataset,
either for the entire dataset or within specified strata. It returns the names or counts of
variables that are found to be non-normally distributed based on a significance level of 0.05.
Before this test, all variables which are desired to be tests, should be defined in explanatory vector.
}
\examples{
\dontrun{

table_vars <- c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
or
table_vars <- iris \%>\%
dplyr::select(-Species) \%>\%
names()
dependent <- "Species"

ag_shapiro(dataset = iris, table_vars = table_vars)
ag_shapiro(dataset = iris, table_vars = table_vars, names = TRUE, message = TRUE)
}


}
\author{
Ali Guner
}
