% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{col_table_header}
\alias{col_table_header}
\title{Constants for package}
\format{
a vector with the colors I need
}
\usage{
col_table_header
}
\description{
This file contains commonly used constants for the package.
}
\keyword{internals}
