% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_colors.R
\name{scale_fill_karadeniz}
\alias{scale_fill_karadeniz}
\title{scale_fill_karadeniz}
\usage{
scale_fill_karadeniz(palette = "Faroz", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{hoose your favourite palette: "Hidirnebi", "Kadirga", "Kayabasi",  "Cal", "Pokut", "Samistal", "Elevit", "Gito" , "Ovit", "Palovit", "Faroz", "Vazil"}

\item{discrete}{TRUE/FALSE}

\item{reverse}{TRUE/FALSE}
}
\description{
to FILL function
}
\examples{
\dontrun{
ggplot(data=iris,mapping=aes(x=Petal.Length,y=Petal.Width, color=Species))+
geom_point() + scale_fill_karadeniz()
}

}
\author{
Ali Guner
}
